refactor(org): Operations becomes Resources; drop Waitlist and Trials tabs
Renames the org dashboard's Operations page to Resources and narrows it to
Documents + Recordings.

Waitlist is dropped because the waitlist feature is being retired outright,
so giving it a tab would be building navigation for something on its way
out. Trials is dropped because a trial IS an appointment — listing it
separately splits one concept across two surfaces, which is the exact
problem this consolidation set out to remove. Both panels are deleted
rather than hidden; nothing else referenced them.

What remains is genuinely one idea — the artefacts a session leaves behind
— so "Operations" (which described a grab-bag of booking-side feeds) is no
longer the right name. The sidebar group and its collapse default are
renamed to match, and the icon moves from ListOrdered to FolderOpen.

Still gated on `operations.read`, unchanged: the permission key names a
grant, not a page, and both surfaces already used it.

Part of #1021
