+++ app/dashboard/organization/[orgId]/resources/RecordingsPanel.tsx
