+++ docs/enterprise/30-programs-and-lifecycle/04-dashboard-pages.md
@@ -72,9 +72,8 @@ The list below is the actual `page.tsx` set under
                                                   summary (host-side only)
 /dashboard/organization/[orgId]/analytics      → rollups (bookings, revenue,
                                                   earnings, wallet burn-down)
-/dashboard/organization/[orgId]/operations     → booking-side feeds, with
-                                                  ?tab=waitlist | trials |
-                                                  documents | recordings
+/dashboard/organization/[orgId]/resources      → session artefacts, with
+                                                  ?tab=documents | recordings
 /dashboard/organization/[orgId]/disputes       → payment-dispute tracker
 /dashboard/organization/[orgId]/reimbursements → (see above)
 /dashboard/organization/[orgId]/audit          → per-org OrgAuditLog (rich filters)
@@ -95,7 +94,7 @@ link into a specific panel still works.
 | Page | Tabs | Why they merged |
 |---|---|---|
 | `/members` | `all`, `learners`, `experts`, `invitations` | `learners` and `experts` were `?role=` queries against the same `/api/organizations/[orgId]/members` endpoint the roster already read. `learners` was additionally capped at `perPage=100` with no pagination. |
-| `/operations` | `waitlist`, `trials`, `documents`, `recordings` | Four read-only `ScopedListTable` views, ~110 lines apiece, with no actions on any of them. |
+| `/resources` | `documents`, `recordings` | Two read-only `ScopedListTable` views, ~110 lines apiece, with no actions on either. Waitlist and Trials were briefly tabs here too; the waitlist feature is being retired, and a trial IS an appointment so it belongs on `/appointments`. |
 | `/settings` | `general`, `sso`, `webhooks`, `scim`, `data-exports` | SSO had no sidebar entry at all and was reachable only from a link inside the settings page. |
 | `/billing` | `invoices`, `wallet` | Unchanged — this one predates the consolidation. |
 
@@ -246,8 +245,8 @@ readable projection of it.
 > `OrganizationPlan` model) is gone. Discovery now reads each per-type
 > plan's `OrgPlanVisibility` directly — see
 > [public pages & discovery](05-public-pages-and-discovery.md). The
-> operations surfaces (`/appointments`, and the `/operations` tabs
-> `waitlist`, `trials`, `documents`, `recordings`) all share the single `operations.read`
+> operations surfaces (`/appointments`, and the `/resources` tabs
+> `documents`, `recordings`) all share the single `operations.read`
 > grant (OWNER, MAINTAINER, MANAGER, SUPPORT) at sidebar, page, and API;
 > `/reimbursements` uses `reimbursements.read` plus the
 > `fundingSource=PERSONAL` structural gate; `/audit` uses `audit.read`
